refactor: extract incremental aggregates module
